+++ tsa-core/src/test/kotlin/org/ton/examples/contracts/ContractsTest.kt
@@ -2,6 +2,7 @@ package org.ton.examples.contracts
 
 import org.ton.examples.checkAtLeastOneStateForAllMethods
 import org.ton.examples.funcCompileAndAnalyzeAllMethods
+import java.math.BigInteger
 import kotlin.io.path.Path
 import kotlin.test.Ignore
 import kotlin.test.Test
@@ -124,7 +125,7 @@ class ContractsTest {
     private fun analyzeContract(
         contractPath: String,
         methodsNumber: Int,
-        methodsBlackList: Set<Int> = setOf(Int.MAX_VALUE)
+        methodsBlackList: Set<BigInteger> = setOf(Int.MAX_VALUE.toBigInteger())
     ) {
         val bytecodeResourcePath = this::class.java.getResource(contractPath)?.path?.let { Path(it) }
             ?: error("Cannot find resource bytecode $contractPath")
